created MNIST classification experiment
I expect this to take a few days to run, so I created a saving mechanism (unlike the other experiments, which only take a matter of minutes). This benchmark, despite its computational cost, is invaluable because it shows a model's true ability to deal with long-term dependencies in a "real world" situation.